+++ src/config/typescript.js
@@ -19,6 +19,8 @@ export default {
       rules: {
         camelcase: 'off',
         'no-empty-function': 'off',
+        // Enabled by eslint:recommended, conflicts with @typescript-eslint/no-extra-semi
+        'no-extra-semi': 'off',
         // This rule is enabled by `eslint-plugin-problems`; we replace it here with
         // `@typescript-eslint/return-await`, which does the same thing as this rule but with better
         // type information.
@@ -26,6 +28,12 @@ export default {
         'no-unused-vars': 'off',
         'no-use-before-define': 'off',
         'no-useless-constructor': 'off',
+
+        '@typescript-eslint/ban-ts-comment': 'error',
+        // Enabled by @typescript-eslint/recommended, but @typescript-eslint/ban-ts-comment is a
+        // more restrictive version of this rule.
+        '@typescript-eslint/ban-ts-ignore': 'off',
+        '@typescript-eslint/default-param-last': 'warn',
         // Enabled by @typescript-eslint/recommended, conflicts with prettier
         '@typescript-eslint/indent': 'off',
         // TODO(ndhoule): Enabled by @typescript-eslint/recommended. Discuss whether or not this is
@@ -35,15 +43,19 @@ export default {
         // but will eventually, accept any props.
         '@typescript-eslint/no-empty-interface': 'off',
         '@typescript-eslint/no-extra-non-null-assertion': 'error',
+        '@typescript-eslint/no-extra-semi': ['error'],
         '@typescript-eslint/no-for-in-array': 'error',
+        '@typescript-eslint/no-implied-eval': 'error',
         // TODO(ndhoule): Enabled by @typescript-eslint/recommended. Discuss enabling it permanently.
         '@typescript-eslint/no-this-alias': 'off',
+        '@typescript-eslint/no-throw-literal': 'error',
         '@typescript-eslint/no-unnecessary-type-assertion': 'error',
         '@typescript-eslint/no-unused-vars': ['error', recommended.rules['no-unused-vars'][1]],
         '@typescript-eslint/no-use-before-define': ['error', {functions: false, typedefs: false}],
         '@typescript-eslint/no-useless-constructor': 'error',
         // This rule is useful, but it duplicates the functionality offered by `import/no-commonjs`.
         '@typescript-eslint/no-var-requires': 'off',
+        '@typescript-eslint/prefer-as-const': 'error',
         '@typescript-eslint/promise-function-async': 'error',
         '@typescript-eslint/restrict-plus-operands': ['error', {checkCompoundAssignments: true}],
         '@typescript-eslint/restrict-template-expressions': 'error',